[SRP] | Try to refactor the CarController from the ProductController
